pyproject.toml
[tool.poetry]
name = "mediasorter-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    {from = "src/python", include = "api"}
]

[tool.poetry.scripts]
mediasorter-server = "api.app:main"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
multimediasorter = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
